py_requirements.txt
numpy==1.18.1
pkg-resources==0.0.0
pyusb==1.0.2
